+++ src/main/java/shadows/gateways/gate/Wave.java
@@ -5,8 +5,6 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-import org.apache.commons.lang3.tuple.Pair;
-
 import com.google.gson.JsonArray;
 import com.google.gson.JsonDeserializationContext;
 import com.google.gson.JsonDeserializer;
@@ -19,16 +17,14 @@ import com.google.gson.reflect.TypeToken;
 
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
-import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.util.GsonHelper;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Mob;
 import net.minecraft.world.entity.MobSpawnType;
@@ -38,12 +34,10 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraftforge.event.ForgeEventFactory;
-import net.minecraftforge.registries.ForgeRegistries;
 import shadows.gateways.GatewayObjects;
 import shadows.gateways.Gateways;
 import shadows.gateways.entity.GatewayEntity;
-import shadows.placebo.json.ItemAdapter;
-import shadows.placebo.json.JsonUtil;
+import shadows.placebo.json.PlaceboJsonReloadListener;
 import shadows.placebo.json.RandomAttributeModifier;
 import shadows.placebo.util.RandomRange;
 
@@ -55,20 +49,19 @@ import shadows.placebo.util.RandomRange;
  * @param maxWaveTime The time the player has to complete this wave.
  * @param setupTime The delay after this wave before the next wave starts.  Ignored if this is the last wave.
  */
-public record Wave(List<Pair<EntityType<?>, CompoundTag>> entities, List<RandomAttributeModifier> modifiers, List<Reward> rewards, int maxWaveTime, int setupTime) {
+public record Wave(List<WaveEntity> entities, List<RandomAttributeModifier> modifiers, List<Reward> rewards, int maxWaveTime, int setupTime) {
 
 	public List<LivingEntity> spawnWave(ServerLevel level, BlockPos pos, GatewayEntity gate) {
 		List<LivingEntity> spawned = new ArrayList<>();
-		for (Pair<EntityType<?>, CompoundTag> toSpawn : entities) {
-			EntityType<?> type = toSpawn.getKey();
-			CompoundTag tag = toSpawn.getValue();
+		for (WaveEntity toSpawn : entities) {
+
 			double spawnRange = gate.getGateway().getSpawnRange();
 
 			int tries = 0;
 			double x = pos.getX() + (-1 + 2 * level.random.nextDouble()) * spawnRange;
 			double y = pos.getY() + level.random.nextInt(3) - 1;
 			double z = pos.getZ() + (-1 + 2 * level.random.nextDouble()) * spawnRange;
-			while (!level.noCollision(type.getAABB(x, y, z)) && tries++ < 7) {
+			while (!level.noCollision(toSpawn.getAABB(x, y, z)) && tries++ < 7) {
 				x = pos.getX() + (level.random.nextDouble() - level.random.nextDouble()) * spawnRange + 0.5D;
 				y = pos.getY() + level.random.nextInt(3 * (int) gate.getGateway().getSize().getScale()) - 1;
 				z = pos.getZ() + (level.random.nextDouble() - level.random.nextDouble()) * spawnRange + 0.5D;
@@ -76,24 +69,24 @@ public record Wave(List<Pair<EntityType<?>, CompoundTag>> entities, List<RandomA
 
 			final double fx = x, fy = y, fz = z;
 
-			if (level.noCollision(type.getAABB(x, y, z))) {
-				Entity entity = type.create(level);
+			if (level.noCollision(toSpawn.getAABB(fx, fy, fz))) {
+				LivingEntity entity = toSpawn.createEntity(level);
 
-				if (!(entity instanceof LivingEntity)) {
+				if (entity == null) {
 					Gateways.LOGGER.error("Gate {} failed to create a living entity during wave {}!", gate.getName().getString(), gate.getWave());
 					continue;
 				}
-				LivingEntity living = (LivingEntity) entity;
 
-				if (tag != null) entity.load(tag);
 				entity.moveTo(fx, fy, fz, level.random.nextFloat() * 360, level.random.nextFloat() * 360);
 
-				modifiers.forEach(m -> m.apply(level.random, living));
-				living.setHealth(living.getMaxHealth());
-				living.addEffect(new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE, 5, 100, true, false));
+				entity.getPassengersAndSelf().filter(e -> e instanceof LivingEntity).map(LivingEntity.class::cast).forEach(e -> {
+					modifiers.forEach(m -> m.apply(level.random, e));
+					e.setHealth(entity.getMaxHealth());
+					e.addEffect(new MobEffectInstance(MobEffects.DAMAGE_RESISTANCE, 5, 100, true, false));
+				});
 
 				if (entity instanceof Mob mob) {
-					if ((tag == null || tag.isEmpty() || tag.getBoolean("ForceFinalizeSpawn")) && !ForgeEventFactory.doSpecialSpawn((Mob) entity, (LevelAccessor) level, (float) entity.getX(), (float) entity.getY(), (float) entity.getZ(), null, MobSpawnType.NATURAL)) {
+					if (toSpawn.shouldFinalizeSpawn() && !ForgeEventFactory.doSpecialSpawn((Mob) entity, (LevelAccessor) level, (float) entity.getX(), (float) entity.getY(), (float) entity.getZ(), null, MobSpawnType.NATURAL)) {
 						mob.finalizeSpawn(level, level.getCurrentDifficultyAt(entity.blockPosition()), MobSpawnType.NATURAL, null, null);
 					}
 					mob.setTarget(gate.getLevel().getNearestPlayer(gate, 12));
@@ -123,10 +116,11 @@ public record Wave(List<Pair<EntityType<?>, CompoundTag>> entities, List<RandomA
 	public JsonObject write() {
 		JsonObject obj = new JsonObject();
 		JsonArray arr = new JsonArray();
-		for (Pair<EntityType<?>, CompoundTag> entity : entities) {
-			JsonObject entityData = new JsonObject();
-			entityData.addProperty("entity", entity.getKey().getRegistryName().toString());
-			if (entity.getValue() != null) entityData.add("nbt", ItemAdapter.ITEM_READER.toJsonTree(entity.getValue()));
+		for (WaveEntity entity : entities) {
+			var s = entity.getSerializer();
+			ResourceLocation id = WaveEntity.SERIALIZERS.inverse().get(s);
+			JsonObject entityData = s.write(entity);
+			entityData.addProperty("type", id.toString());
 			arr.add(entityData);
 		}
 		obj.add("entities", arr);
@@ -139,12 +133,12 @@ public record Wave(List<Pair<EntityType<?>, CompoundTag>> entities, List<RandomA
 
 	public static Wave read(JsonObject obj) {
 		JsonArray entities = obj.get("entities").getAsJsonArray();
-		List<Pair<EntityType<?>, CompoundTag>> entityList = new ArrayList<>();
+		List<WaveEntity> entityList = new ArrayList<>();
 		for (JsonElement e : entities) {
 			JsonObject entity = e.getAsJsonObject();
-			EntityType<?> type = JsonUtil.getRegistryObject(entity, "entity", ForgeRegistries.ENTITIES);
-			CompoundTag nbt = entity.has("nbt") ? ItemAdapter.ITEM_READER.fromJson(entity.get("nbt"), CompoundTag.class) : null;
-			entityList.add(Pair.of(type, nbt));
+			ResourceLocation id = entity.has("type") ? new ResourceLocation(entity.get("type").getAsString()) : PlaceboJsonReloadListener.DEFAULT;
+			var s = WaveEntity.SERIALIZERS.get(id);
+			entityList.add(s.read(entity));
 		}
 		List<RandomAttributeModifier> modifiers = Gateway.GSON.fromJson(obj.get("modifiers"), new TypeToken<List<RandomAttributeModifier>>() {
 		}.getType());
@@ -159,8 +153,11 @@ public record Wave(List<Pair<EntityType<?>, CompoundTag>> entities, List<RandomA
 
 	public void write(FriendlyByteBuf buf) {
 		buf.writeVarInt(entities.size());
-		for (Pair<EntityType<?>, CompoundTag> entity : entities) {
-			buf.writeRegistryId(entity.getKey());
+		for (WaveEntity entity : entities) {
+			var s = entity.getSerializer();
+			ResourceLocation id = WaveEntity.SERIALIZERS.inverse().get(s);
+			buf.writeResourceLocation(id);
+			s.write(entity, buf);
 		}
 		buf.writeVarInt(modifiers.size());
 		modifiers.forEach(m -> {
@@ -176,9 +173,11 @@ public record Wave(List<Pair<EntityType<?>, CompoundTag>> entities, List<RandomA
 
 	public static Wave read(FriendlyByteBuf buf) {
 		int size = buf.readVarInt();
-		List<Pair<EntityType<?>, CompoundTag>> entities = new ArrayList<>(size);
+		List<WaveEntity> entities = new ArrayList<>(size);
 		for (int i = 0; i < size; i++) {
-			entities.add(Pair.of(buf.readRegistryIdSafe(EntityType.class), null));
+			ResourceLocation id = buf.readResourceLocation();
+			var s = WaveEntity.SERIALIZERS.get(id);
+			entities.add(s.read(buf));
 		}
 		size = buf.readVarInt();
 		List<RandomAttributeModifier> modifiers = new ArrayList<>(size);